refactor: share repeated theme generation logic
